requirements.txt
python-dotenv
openai
datasets
ragas
deepeval
pinecone
pinecone[grpc]
